Implementing adapters hexagon
This commit contains WAY too much code, and implements the entire adapters' hexagon. As a result, a follow-up commit will include tests coverage for both the adapter hexagon, and changes in the application/domain hexagon.

However, at this point in time the application is fully functional via a CLI interaction 🎉